# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: USS OPS - Ultimate Stain & Seal Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/disclaimer:
    patch:
      operationId: updateInvoiceDisclaimer
      tags: [invoices]
      summary: Set soft or hard disclaimer mode on invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceDisclaimerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ─── PAYMENT RECORDS ─────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [invoices]
      summary: List payment records
      parameters:
        - name: invoiceId
          in: query
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [invoices]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [invoices]
      summary: Get payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [invoices]
      summary: Update payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─── JOBS / SCHEDULING ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        customerId:
          type: integer
        invoiceNumber:
          type: string
        invoiceType:
          type: string
          enum: [standard, traditional_soft, traditional_hard]
        status:
          type: string
          enum: [draft, sent, viewed, approved, partial, paid, overdue, void]
        disclaimerMode:
          type: string
          enum: [none, soft, hard]
        subtotal:
          type: number
        taxRate:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        disclaimerText:
          type: ["string", "null"]
        customerSignature:
          type: ["string", "null"]
        signedAt:
          type: ["string", "null"]
        glideId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, customerId, invoiceNumber, invoiceType, status, disclaimerMode, subtotal, totalAmount, paidAmount, balanceDue, createdAt, updatedAt]
    InvoiceDetail:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        customerId:
          type: integer
        invoiceNumber:
          type: string
        invoiceType:
          type: string
        status:
          type: string
        disclaimerMode:
          type: string
        subtotal:
          type: number
        taxRate:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        disclaimerText:
          type: ["string", "null"]
        customerSignature:
          type: ["string", "null"]
        signedAt:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        glideId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, customerId, invoiceNumber, invoiceType, status, disclaimerMode, subtotal, totalAmount, paidAmount, balanceDue, lineItems, payments, createdAt, updatedAt]
    InvoiceLineItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        category:
          type: ["string", "null"]
      required: [id, invoiceId, description, quantity, unitPrice, lineTotal]
    CreateInvoiceBody:
      type: object
      properties:
        projectId:
          type: integer
        customerId:
          type: integer
        invoiceType:
          type: string
          enum: [standard, traditional_soft, traditional_hard]
        disclaimerMode:
          type: string
          enum: [none, soft, hard]
        subtotal:
          type: number
        taxRate:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        disclaimerText:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: number
              unitPrice:
                type: number
              category:
                type: ["string", "null"]
            required: [description, quantity, unitPrice]
      required: [projectId, customerId, invoiceType, disclaimerMode, subtotal]
    UpdateInvoiceBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        disclaimerMode:
          type: ["string", "null"]
        subtotal:
          type: ["number", "null"]
        taxRate:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        disclaimerText:
          type: ["string", "null"]
        customerSignature:
          type: ["string", "null"]
    InvoiceDisclaimerBody:
      type: object
      properties:
        disclaimerMode:
          type: string
          enum: [none, soft, hard]
        disclaimerText:
          type: ["string", "null"]
      required: [disclaimerMode]

    # ─── PAYMENTS ─────────────────────────────────────────────────────────────
    Payment:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        customerId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, check, credit_card, zelle, venmo, ach, other]
        paymentDate:
          type: string
        checkNumber:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        recordedById:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, invoiceId, customerId, amount, paymentMethod, paymentDate, createdAt, updatedAt]
    CreatePaymentBody:
      type: object
      properties:
        invoiceId:
          type: integer
        customerId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, check, credit_card, zelle, venmo, ach, other]
        paymentDate:
          type: string
        checkNumber:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        recordedById:
          type: ["integer", "null"]
      required: [invoiceId, customerId, amount, paymentMethod, paymentDate]
    UpdatePaymentBody:
      type: object
      properties:
        amount:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
        checkNumber:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── JOBS ─────────────────────────────────────────────────────────────────
...